Update frontend to plain javascript
