+++ frontend/.prettierignore
@@ -1,4 +1,5 @@
-node_modules
 # Ignore artifacts:
 build
-Coverage
+coverage
+node_modules
+public
