+++ frontend/.prettierrc.json
@@ -1,9 +1,9 @@
 {
-  "useTabs": false,
-  "tabWidth": 2,
-  "printWidth": 120,
-  "bracketSpacing": false,
-  "jsxBracketSameLine": false,
-  "singleQuote": true,
-  "trailingComma": "all"
+	"tabWidth": 2,
+	"useTabs": true,
+	"printWidth": 80,
+	"semi": true,
+	"trailingComma": "es5",
+	"singleQuote": true,
+	"endOfLine": "lf"
 }
