--- frontend/src/App.tsx
@@ -1,32 +1,0 @@
-import React from "react";
-import logo from "./assets/logo.svg";
-import {Routes, Route, BrowserRouter} from "react-router-dom";
-import "./App.css";
-
-function App() {
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<LandingPage />} />
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-function LandingPage(): React.ReactElement {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
