--- frontend/src/redux/reducers/index.ts
@@ -1,7 +1,0 @@
-import { combineReducers } from "redux";
-
-const rootReducer = combineReducers({
-  // Add your individual reducers here
-});
-
-export default rootReducer;
