--- frontend/src/redux/saga/index.ts
@@ -1,5 +1,0 @@
-import { all } from "redux-saga/effects";
-
-export default function* rootSaga() {
-  yield all([]);
-}
