+++ frontend/src/redux/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit';
+import rootReducer from './reducers';
+
+const store = configureStore({
+	reducer: rootReducer,
+	middleware: (getDefaultMiddleware) => getDefaultMiddleware(),
+	devTools: process.env.NODE_ENV !== 'production',
+});
+
+export default store;
