--- frontend/src/sample.env
@@ -1,1 +1,0 @@
-NODE_ENV=
